requirements.txt
requests>=2.28.0
lxml>=4.9.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Database tools
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
sqlite3

# Data analysis and visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0

# Additional utilities
python-dotenv>=1.0.0
tqdm>=4.64.0
